% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HapDipKinship.R
\name{dipking}
\alias{dipking}
\title{Diploid-diploid kinship estimation}
\usage{
dipking(dip1, dip2, dipref = NA)
}
\arguments{
\item{dip1, dip2}{Vectors encoding the genotype of diploid individuals (0,1,2).}

\item{dipref}{A vector encoding the genotype of the reference diploid individual (0,1,2).
Perform exKING-robust if none provided.}
}
\description{
Estimate kinship for a pair of diploid individuals.
}
\examples{
##Calculate kinship of a pair of full siblings
dip<-rbinom(1000,2,0.5)
hap<-rbinom(1000,1,0.5)
### without reference(exKING-robust)
dipking(anastomosis(meiosis(dip),hap),anastomosis(meiosis(dip),hap1))
### with reference(KIMGENS)
dipking(anastomosis(meiosis(dip),hap),anastomosis(meiosis(dip),hap1),dipref=dip1)
}
