% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopSims.R
\name{ExtractPairsOfInterest}
\alias{ExtractPairsOfInterest}
\title{Extract pairs of interest}
\usage{
ExtractPairsOfInterest(kinships, PairsOfInterest)
}
\arguments{
\item{kinships}{A data frame with at least two columns: ind1 and ind2}

\item{PairsOfInterest}{A data frame describing focal relationships.
All ids in ind1 and ind2 should be found in kinships (the order is not important).
Report all if none provided. See \link{PairsOfInterest}.}
}
\value{
Reduced data frame of Kinship.
}
\description{
A helper function to extract kinships of specific pairs of individuals.
}
\examples{
ancestrygenomatrix<-PopulationSim(1000,c(0.05,0.15,0.25))
ancestry<-matrix(c(6,2,0.3,2,6,0.3),nrow=3)
kins<-HapdipPedigreeSim(ancestrygenomatrix,pedigree,ancestry)
ExtractPairsOfInterest(kins,PairsOfInterest)
}
