% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HapDipKinship.R
\name{hapdipking}
\alias{hapdipking}
\title{Haploid-diploid kinship estimation}
\usage{
hapdipking(dip, hap, dipref = NA)
}
\arguments{
\item{dip}{A vector encoding the genotype of a diploid individual (0,1,2).}

\item{hap}{A vector encoding the genotype of a haploid individual (0,1).}

\item{dipref}{A vector encoding the genotype of the reference diploid individual (0,1,2).
Perform exKING-robust hap-dip kinship estimation if none provided.}
}
\description{
Estimate kinship between a haploid and a diploid individuals.
}
\examples{
##Calculate kinship of a pair of full siblings
dip<-rbinom(1000,2,0.5)
hap<-rbinom(1000,1,0.5)
### without reference(exKING-robust)
hapdipking(anastomosis(meiosis(dip),hap),meiosis(dip))
### with reference(KIMGENS)
hapdipking(anastomosis(meiosis(dip),hap),meiosis(dip),dipref=dip)
}
