% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HapDipKinship.R
\name{hapking}
\alias{hapking}
\title{Haploid-haploid kinship estimation}
\usage{
hapking(hap1, hap2, dipref)
}
\arguments{
\item{hap1, hap2}{Vectors encoding the genotype of haploid individuals (0,1).}

\item{dipref}{A vector encoding the genotype of the reference diploid individual (0,1,2).
Required!}
}
\description{
Estimate kinship for a pair of haploid individuals.
}
\examples{
##Calculate kinship of a pair of full siblings
dip<-rbinom(1000,2,0.5)
### No exKING-robust
hapking(meiosis(dip),meiosis(dip),dipref=dip)
}
