% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HapDipKinship.R
\name{kinship}
\alias{kinship}
\title{Kinship}
\usage{
kinship(
  genotypematrix,
  ploidy = NA,
  skipRelBased = F,
  RelBasedKinshipThreshold = 0.1
)
}
\arguments{
\item{genotypematrix}{A matrix encoding genotypes with columns of individuals and rows of SNP sites. (0=homozygotic reference, 1=heterozygotic, 2=homozygotic alternative).}

\item{ploidy}{A data frame with two columns.
First column is id (same as column names of genotype matrix).
Second column is ploidy (1 or 2).
Guess ploidy if not provided.}

\item{skipRelBased}{Binary variable for skipping relative-based kinship estimates or not.
F if not provided. See 'Details.'}

\item{RelBasedKinshipThreshold}{A kinship threshold for defining "relatives" in relative-based kinship. 0.1 if not provided. See 'Details.'}
}
\value{
A dataframe encoding kinship and IBS0 for every pair of individuals
}
\description{
Estimates the kinships for all combinations of pairs of individuals.
}
\details{
This functions takes two step:
The first step use KING-robust (Manichaikul et al. 2010) and XXX (new algorithm) to estimate kinship between two diploids and kinship between a haploid and a diploid.
The second step use individuals related to the two individuals of interest as "references" to estimate kinship.
The median of kinships from relatives of either individuals are averaged to estimate the correct kinship
Relatives are defined by if the kinship estimates from the first step are larger than RelBasedKinshipThreshold.
Skipping this step will skip haploid-haploid kinship estimate.
}
\examples{
tbd
}
