% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopSims.R
\name{meiosis}
\alias{meiosis}
\title{Meiosis}
\usage{
meiosis(dip)
}
\arguments{
\item{dip}{A vector encoding the genotype of a diploid individual (0,1,2).}
}
\value{
A vector encoding the genotype of a haploid individual (0,1).
}
\description{
Take a diploid individual and generate gamete
}
\examples{
dip<-sample(c(0:2),10,replace=T)
meiosis(dip)
}
