% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pedigree}
\alias{pedigree}
\title{Pedigrees}
\format{
A data frame with 42 rows and 4 variables:
\describe{
\item{id}{individuals' id.}
\item{father}{individual's father.}
\item{mother}{individual's mother}
\item{sex}{individuals' sex.}
}
}
\source{
\url{https://www.github.com/YenWenWang/HapDipKinship}
}
\usage{
pedigree
}
\description{
An example data set with parents of individuals.
}
\keyword{datasets}
