% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RelOfInterest}
\alias{RelOfInterest}
\title{Relatedness of interest}
\format{
A data frame with 164 rows and 4 variables:
\describe{
\item{ind1}{individual 1.}
\item{ind1}{individual 2.}
\item{relationship}{manually coded relationship.}
\item{pairtype}{ploidy of ind1 and ind2.}
}
}
\source{
\url{https://www.github.com/YenWenWang/HapDipKinship}
}
\usage{
RelOfInterest
}
\description{
A dataset denoting which relatives are of interest (for simulation).
}
\keyword{datasets}
