% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopSims.R
\name{anastomosis}
\alias{anastomosis}
\title{Anastomosis}
\usage{
anastomosis(hap1, hap2)
}
\arguments{
\item{hap1}{A vector encoding the genotype of a haploid individual (0,1).}

\item{hap2}{A vector encoding the genotype of a haploid individual (0,1).}
}
\value{
A vector encoding the genotype of a diploid individual (0,1,2).
}
\description{
Combine two gametes into a diploid individual
}
\examples{
hap1<-sample(c(0:1),10,replace=T)
hap2<-sample(c(0:1),10,replace=T)
anastomosis(hap1,hap2,replace=T)
}
