% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageDescription.R
\docType{package}
\name{HapDipKinship-package}
\alias{HapDipKinship}
\alias{HapDipKinship-package}
\title{HapDipKinship: HapDipKinship: Kinship estimation in haploid-diploid populations}
\description{
This package is for estimating kinship and simulating pedigrees in a haploid-diploid mixed dataset.
}
\details{
New algorithms are developed to estimate kinships among haploid and diploid individuals.
They are largely inspired by KING-robust (Manichaikul, et al. 2010).
The package also offers some simple simulation to test the algorithms.
}
\references{
Wang et al. in prep. Kinship algorithm for haploid-diploid mixed populations.
Manichaikul, et al. 2010. Robust relationship inference in genome-wide association studies. Bioinformatics 26(22): 2867–2873.
}
\author{
\strong{Maintainer}: Yen-Wen Wang \email{ywang883@wisc.edu} (\href{https://orcid.org/0000-0002-6851-1646}{ORCID})

}
\keyword{internal}
