% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopSims.R
\name{HapdipPedigreeSim}
\alias{HapdipPedigreeSim}
\title{Simulate Pedigree}
\usage{
HapdipPedigreeSim(ancestrygenomatrix, pedigree, ancestry, ancestryprop = NA)
}
\arguments{
\item{ancestrygenomatrix}{A matrix of allele frequencies in ancestral subpopulations.}

\item{pedigree}{A data frame describing the simulation of a pedigree. See \link{pedigree}.}

\item{ancestry}{A matrix including dirichlet params for the admixture of ancestral subpopulations.
Rows denote ancestral subpopulations. Columns denote different admixture schemes.}

\item{ancestryprop}{A vector with the length of number of column of ancestry
denoting the probability of drawing individuals from each admixed populations.
Average sampling if none give.}
}
\value{
Kinships of PairsOfInterest from simulated populations.
}
\description{
Simulate a haplodiploidy population pedigree and calculate kinships from a few subpopulations
}
\examples{
ancestrygenomatrix<-PopulationSim(1000,c(0.05,0.15,0.25))
ancestry<-matrix(c(6,2,0.3,2,6,0.3),nrow=3)
HapdipPedigreeSim(ancestrygenomatrix,pedigree,ancestry)
}
\seealso{
\code{\link[=PopulationSim]{PopulationSim()}},
\code{\link[=kinship]{kinship()}}
}
