% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HapDipKinship.R
\name{kinship}
\alias{kinship}
\title{Kinship}
\usage{
kinship(
  genotypematrix,
  ploidy,
  skipRelBased = F,
  RelBasedKinshipThreshold = 0.1
)
}
\arguments{
\item{genotypematrix}{A matrix encoding genotypes with columns of individuals and rows of SNP sites. (0=homozygotic reference, 1=heterozygotic, 2=homozygotic alternative).}

\item{ploidy}{A data frame with two columns. First column is id (same as column names of genotype matrix). Second column is ploidy (1 or 2). Guess ploidy if not provided.}

\item{skipRelBased}{Binary variable for skipping relative-based kinship estimates or not. F if not provided. See 'Details.'}

\item{RelBasedKinshipThreshold}{A kinship threshold for defining "relatives" in relative-based kinship. 0.1 if not provided. See 'Details.'}
}
\value{
A vector encoding the genotype of a haploid individual (0,1).
}
\description{
Estimates the kinships for all combinations of pairs of individuals.
}
\examples{
dip<-sample(c(0:2),10,replace=T)
meiosis(dip)
}
