% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopSims.R
\name{PopulationSim}
\alias{PopulationSim}
\title{Simulate Population}
\usage{
PopulationSim(sites, thetak, pthresh = 0.1, beta = 1)
}
\arguments{
\item{sites}{Number of sites simulated.}

\item{thetak}{An x-length vector of Fsts for x ancestral subpopulations (ancestral subpopulations vs total ancestral population).}

\item{pthresh}{A threshold to avoid low frequency SNPs. 0.1 if not provided.}

\item{beta}{Shape parameter for beta distribution for simulating original allele frequency. 1 if not provided.}
}
\value{
A matrix of allele frequencies in ancestral subpopulations.
}
\description{
Simulate population with substructure using Balding-Nichols model.
}
\examples{
PopulationSim(1000,c(0.05,0.15,0.25))
}
\seealso{
\code{\link[=HapdipPedigreeSim]{HapdipPedigreeSim()}}
}
